Remove check for PM context in hook_privmsg()
Removed the call to get_current_context_type in hook_privmsg, which checks for a
current PM context, and the return if one is not found. This appears to have been
mistakenly copied from hook_outgoing. The original code resulted in non-decrypted
messages being received if the PM/query tab did not have focus.

Closes #3